+++ src/pages/dashboard/dynamic/index.vue
