+++ src/pages/dashboard/index.vue
