--- src/pages/home/index.vue
@@ -1,8 +1,0 @@
-<template>
-  <div>
-    <HelloWorld msg="Hello Vue 3 + TypeScript + Vite" />
-  </div>
-</template>
-<script lang="ts" setup>
-import HelloWorld from '../../components/HelloWorld.vue'
-</script>
